+++ src/Creational/Factories/Abstract/Interface/Product.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Codenip\Creational\Factories\Abstract\Interface;
+
+interface Product {}