+++ src/Creational/Factories/Abstract/Interface/ProductFactory.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Codenip\Creational\Factories\Abstract\Interface;
+
+interface ProductFactory
+{
+    public function createProduct(): Product;
+
+    public function createSparePartProduct(): SparePartProduct;
+}