+++ src/Creational/Factories/Abstract/SparePartProduct/UsedSparePartProduct.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Codenip\Creational\Factories\Abstract\SparePartProduct;
+
+use Codenip\Creational\Factories\Abstract\Interface\SparePartProduct;
+
+class UsedSparePartProduct implements SparePartProduct {}